+++ modules/core/src/model/dynamic-form-control-relation.model.ts
@@ -17,5 +17,5 @@ export interface DynamicFormControlRelationGroup {
 
     action: string;
     connective?: string;
-    when: Array<DynamicFormControlRelation>;
+    when: DynamicFormControlRelation[];
 }