+++ webpack.config.js
@@ -14,10 +14,11 @@ module.exports = {
         "ui-material": "./@ng2-dynamic-forms/ui-material/index.js",
         "ui-primeng": "./@ng2-dynamic-forms/ui-primeng/index.js"
     },
+    /*
     resolve: {
         extensions: ["", ".js"]
     },
-
+    */
     externals: {
 
         "@angular/common": {
@@ -38,29 +39,11 @@ module.exports = {
             commonjs2: "@angular/forms",
             root: ["ng", "forms"]
         },
-        "@angular2-material/checkbox": {
-            amd: "@angular2-material/checkbox",
-            commonjs: "@angular2-material/checkbox",
-            commonjs2: "@angular2-material/checkbox",
-            root: ["md", "checkbox"]
-        },
-        "@angular2-material/core": {
-            amd: "@angular2-material/core",
-            commonjs: "@angular2-material/core",
-            commonjs2: "@angular2-material/core",
-            root: ["md", "core"]
-        },
-        "@angular2-material/input": {
-            amd: "@angular2-material/input",
-            commonjs: "@angular2-material/input",
-            commonjs2: "@angular2-material/input",
-            root: ["md", "input"]
-        },
-        "@angular2-material/radio": {
-            amd: "@angular2-material/radio",
-            commonjs: "@angular2-material/radio",
-            commonjs2: "@angular2-material/radio",
-            root: ["md", "radio"]
+        "@angular/material": {
+            amd: "@angular/material",
+            commonjs: "@angular/material",
+            commonjs2: "@angular/material",
+            root: ["ng", "material"]
         },
         "@ng2-dynamic-forms/core": {
             amd: "@ng2-dynamic-forms/core",
@@ -75,19 +58,19 @@ module.exports = {
         "primeng/components/radiobutton/radiobutton": true,
         "primeng/components/spinner/spinner": true
     },
-
+    /*
     plugins: [
         new webpack.BannerPlugin(
             `${pkg.name} ${pkg.version} ${dateformat(now, "UTC:yyyy-mm-dd HH:MM")} UTC ${license}`
         )
         // new webpack.optimize.UglifyJsPlugin({})
     ],
-
+    */
     output: {
         filename: "[name].umd.js",
         library: ["dynamicForms", "[name]"],
         libraryTarget: "umd",
-        path: "./@ng2-dynamic-forms/",
+        path: "./@ng2-dynamic-forms/test",
         umdNamedDefine: true
     }
 };